requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-fuzzy
pyet
pyyaml
pyarrow
google-cloud-bigquery
google-cloud-bigquery-storage
python-dotenv
requests